Use Apache Ant like a real Java dev!
